+++ node/resolvers/Queries/Organizations.ts
@@ -1,4 +1,7 @@
 import {
+  COST_CENTER_DATA_ENTITY,
+  COST_CENTER_FIELDS,
+  COST_CENTER_SCHEMA_VERSION,
   ORGANIZATION_DATA_ENTITY,
   ORGANIZATION_FIELDS,
   ORGANIZATION_REQUEST_DATA_ENTITY,
@@ -6,7 +9,7 @@ import {
   ORGANIZATION_REQUEST_SCHEMA_VERSION,
   ORGANIZATION_SCHEMA_VERSION,
 } from '../../mdSchema'
-import type { Organization } from '../../typings'
+import type { CostCenter, Organization } from '../../typings'
 import GraphQLError, { getErrorMessage } from '../../utils/GraphQLError'
 import checkConfig from '../config'
 
@@ -112,6 +115,7 @@ const Organizations = {
     _: void,
     {
       status,
+      document,
       search,
       page,
       pageSize,
@@ -119,6 +123,7 @@ const Organizations = {
       sortedBy,
     }: {
       status: string[]
+      document: string | null
       search: string
       page: number
       pageSize: number
@@ -136,6 +141,62 @@ const Organizations = {
     await checkConfig(ctx)
 
     const whereArray = getWhereByStatus({ status })
+
+    if (document) {
+      const documentFormatted = formatDocument(document)
+
+      try {
+        const where = `businessDocument="*${documentFormatted}*"`
+        const costCenterData =
+          await masterdata.searchDocumentsWithPaginationInfo({
+            dataEntity: COST_CENTER_DATA_ENTITY,
+            fields: COST_CENTER_FIELDS,
+            pagination: { page, pageSize },
+            schema: COST_CENTER_SCHEMA_VERSION,
+            sort: `${sortedBy} ${sortOrder}`,
+            ...(where && { where }),
+          })
+
+        const data = costCenterData.data as CostCenter[]
+
+        const getOrganizationFromCostCenter = async (
+          costCenter: CostCenter
+        ) => {
+          const organizations: Organization[] =
+            await masterdata.searchDocuments({
+              dataEntity: ORGANIZATION_DATA_ENTITY,
+              fields: ORGANIZATION_FIELDS,
+              pagination: { page: 1, pageSize: 25 },
+              schema: ORGANIZATION_SCHEMA_VERSION,
+              sort: `${sortedBy} ${sortOrder}`,
+              where: `id="${costCenter.organization}"`,
+            })
+
+          if (organizations.length === 0) {
+            throw new Error('No organization found')
+          }
+
+          if (organizations.length > 1) {
+            throw new Error('More than one organization found')
+          }
+
+          return organizations[0]
+        }
+
+        const orgs = await Promise.all(data.map(getOrganizationFromCostCenter))
+
+        return {
+          data: orgs,
+          pagination: costCenterData.pagination,
+        }
+      } catch (error) {
+        logger.error({
+          error,
+          message: 'getCostCenters-error',
+        })
+        throw new GraphQLError(getErrorMessage(error))
+      }
+    }
 
     if (search) {
       whereArray.push(`name="*${search}*"`)
@@ -406,6 +467,30 @@ const Organizations = {
       throw new GraphQLError(getErrorMessage(error))
     }
   },
+}
+
+function formatDocument(document: string): string {
+  const documentFormatRemoved = document.replace(/[^0-9]/gu, '')
+
+  const documentArray = documentFormatRemoved.split('')
+
+  const formattedDocument = documentArray.map((char, index) => {
+    if (index === 2 || index === 5) {
+      return `.${char}`
+    }
+
+    if (index === 8) {
+      return `/${char}`
+    }
+
+    if (index === 12) {
+      return `-${char}`
+    }
+
+    return char
+  })
+
+  return formattedDocument.join('')
 }
 
 export default Organizations