+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add `document` field to `getOrganizations` query
+
 ## [0.61.0] - 2024-10-16
 
 ### Added