+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Avoid calls to checkUserPermissions when session data is not available
+
 ## [0.61.0] - 2024-10-16
 
 ### Added